move pallet bags list in separate file
